refactored utils.py

+++ libs/utils.py
@@ -5,12 +5,13 @@ from pathlib import Path
 from typing import Optional
 
 
-def recursive_rmdir(directory: str | Path) -> None:
+def recursive_rmdir(directory: str | Path, remove_root: bool = True) -> None:
     """
     Function for recursively clearing a directory (removing all nested files and dirs)
 
     Args:
         directory: path to the directory that will be cleared
+        remove_root: True if needed to delete the root directory too, else False
     """
     try:
         path = Path(directory)
@@ -20,7 +21,8 @@ def recursive_rmdir(directory: str | Path) -> None:
                     entry.unlink()
                 else:
                     recursive_rmdir(entry)
-        path.rmdir()
+        if remove_root:
+            path.rmdir()
     except PermissionError as e:
         print(f"Insufficient rights to delete. Error message: {e}")
     except FileNotFoundError:
@@ -34,18 +36,7 @@ def clear_dir(directory: str | Path) -> None:
     Args:
         directory: path to the directory that will be cleared
     """
-    try:
-        path = Path(directory)
-        if path.is_dir():
-            for entry in path.iterdir():
-                if entry.is_file():
-                    entry.unlink()
-                else:
-                    recursive_rmdir(entry)
-    except PermissionError as e:
-        print(f"Insufficient rights to delete. Error message: {e}")
-    except FileNotFoundError:
-        print(f"File not found: {directory}")
+    recursive_rmdir(directory, False)
 
 
 def clear_or_create_dir(directory: str | Path) -> None:
@@ -67,9 +58,42 @@ def clear_or_create_dir(directory: str | Path) -> None:
         print(f"File not found: {directory}")
 
 
-def format_xlsx(writer: pd.ExcelWriter, df: pd.DataFrame, alignments: str | None = None,
-                sheet_name: str = 'Sheet1', font_size: Optional[int] = None, border_width: Optional[int] = None,
-                border_color: Optional[str] = None, cell_height: int = 20) -> pd.ExcelWriter:
+def crop_image_white_margins(old_filename: str | Path, xpadding: int = 15, ypadding: int = 15,
+                             new_filename: str | Path | None = None) -> None:
+    """
+    Function for cropping images with graphs (white margins at the edges are cut off).
+    If a new filename is not passed, the original file will be overwritten
+
+    Args:
+        old_filename: number of bytes
+        xpadding: horizontal padding
+        ypadding: vertical padding
+        new_filename: path to the new (cropped) image
+    """
+    try:
+        img = cv2.imread(str(old_filename))
+        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+        gray = 255 * (gray < 128).astype(np.uint8)
+        cords = cv2.findNonZero(gray)
+        x, y, w, h = cv2.boundingRect(cords)
+        rect = img[y - ypadding: y + h + 2 * ypadding, x - xpadding: x + w + 2 * xpadding]
+        is_success, im_buf_arr = cv2.imencode(".png", rect)
+        if new_filename is None:
+            Path(old_filename).unlink()
+            im_buf_arr.tofile(old_filename)
+        else:
+            im_buf_arr.tofile(new_filename)
+    except FileNotFoundError as e:
+        print(f"File not found error: {e}")
+    except ValueError as e:
+        print(f"Image reading error: {e}")
+    except IOError as e:
+        print(f"IO error: {e}")
+
+
+def format_xlsx(writer: pd.ExcelWriter, df: pd.DataFrame, sheet_name: str = 'Sheet1',
+                alignments: str | None = None, font_size: int | None = None, border_width: int | None = None,
+                border_color: str | None = None, cell_height: int = 20) -> pd.ExcelWriter:
     """
     Function for formatting an object of XlsxWriter type.
     Allows to set alignment for each column and adjust cells height.
@@ -77,8 +101,8 @@ def format_xlsx(writer: pd.ExcelWriter, df: pd.DataFrame, alignments: str | None
     Args:
         writer: object of XlsxWriter type
         df: pandas dataframe with data
-        alignments: string indicating columns alignments (r, l, c, j), default is left alignment for all columns
         sheet_name: name of the sheet to be formatted
+        alignments: string indicating columns alignments (r, l, c, j), default is left alignment for all columns
         font_size: font size for all cells
         border_width: border width for all cells
         border_color: border color for all cells
@@ -105,35 +129,9 @@ def format_xlsx(writer: pd.ExcelWriter, df: pd.DataFrame, alignments: str | None
             cell_format.set_border_color(border_color)
         worksheet.set_column(col_index, col_index, col_width, cell_format)
     # set cells height
-    for i in range(df.shape[0] + 1):
+    for i in range(len(df) + 1):
         worksheet.set_row(i, cell_height)
     return writer
-
-
-def crop_image_white_margins(old_filename: str | Path, xpadding: int = 15, ypadding: int = 15,
-                             new_filename: str | Path | None = None) -> None:
-    """
-    Function for cropping images with graphs (white margins at the edges are cut off).
-    If a new filename is not passed, the original file will be overwritten
-
-    Args:
-        old_filename: number of bytes
-        xpadding: horizontal padding
-        ypadding: vertical padding
-        new_filename: path to the new (cropped) image
-    """
-    img = cv2.imread(old_filename)
-    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-    gray = 255 * (gray < 128).astype(np.uint8)
-    cords = cv2.findNonZero(gray)
-    x, y, w, h = cv2.boundingRect(cords)
-    rect = img[y - ypadding: y + h + 2 * ypadding, x - xpadding: x + w + 2 * xpadding]
-    is_success, im_buf_arr = cv2.imencode(".png", rect)
-    if new_filename is None:
-        Path(old_filename).unlink()
-        im_buf_arr.tofile(old_filename)
-    else:
-        im_buf_arr.tofile(new_filename)
 
 
 def get_entity_with_case(x, word_forms: tuple[str, str, str] = ('день', 'дня', 'дней')) -> str:
